Ubus: use mac from ubus system board
Due switching to the new DSA driver for xrx200 targets from lantiq,
the use of the eth0 MAC via the sysfs interface is no longer valid.

Also for other targets, 'eth0' is not always the correct interface for the
first assigned MAC of the device.

It is therefore better to read the MAC value via the ubus system board call,
which returns the first valid MAC for this device.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>