rsu-connector: add missing newline at end of file
When saving the file with a linux editor the missing new line is added.
The git programme, after saving the file, always shows
'No newline at end of file' for a `git diff`.

To make the editor happy, the commit adds the missing new line at the
end of file to every file where it is missing.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>