IotHubClientWrapper: add status file output
This change is intended to make the internal connection state to the iothub
accessible for possible further processing outside of the rsu connector.

Therefore, the connection state is saved in a file with every callback
call to the iothub.

There is a new configuration option 'HUB_StatusFile' for the config
file. This option contains a path to a file in which the internal state
is stored as json. This file should be stored in tmp, as it is written
with every callback call. The directory must exist otherwise the file
cannot be created.

The file could have one of the following states:

{ "status": "OK" }
{ "status": "EXPIRED_SAS_TOKEN" }
{ "status": "DEVICE_DISABLED" }
{ "status": "BAD_CREDENTIAL" }
{ "status": "RETRY_EXPIRED" }
{ "status": "NO_NETWORK" }
{ "status": "COMMUNICATION_ERROR" }
{ "status": "OK" }
{ "status": "NO_PING_RESPONSE" }
{ "status": "UNKNOWN" }

Signed-off-by: Florian Eckert <fe@dev.tdt.de>