UBus: change openvpn extension from ovpn to conf
In OpenWrt, the native openvpn OpenVPN configurations are stored under
'/etc/openvpn' with the extention '.conf'. Until now, it was only
possible to start an OpenVPN connection if it also had a uci
configuration section. With the latest change in the OpenVPN start
script, it is no longer necessary to create a uci section for the VPN
connection.

In order for this to work without the uci section, the configuration
must now be saved with the extension '.conf'. This commit replaces the
extension from '.ovpn' to '.conf'.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>