Ubus/Ubus: fix VPNConnectionActive check
When checking the network adapters in the 'GetNetworkDetails' function, an
existing POINT-TO-POINT connection is assumed to be an OpenVPN connection.
However, this is 'not' correct. There are also 'DSL' and WWAN interfaces on
the system, which are also a PONIT-TO-POINT interfaces.

To solve the problem, the system now checks whether the name of the
interface is 'tun100'. This is the network adapter that only exists if
the RSF OpenVpn connection also exists.

This is only a workaround to fix the current problems. Just because the
interface exists does not mean that an OpenVPN connection is full established.
To be correct, the OpenVPN event that is generated when a connection is
successfully established must be evaluated. There is an 'ubus' backend in the
system that provides this information 'ubus call openvpn statu'.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>